Retrieving data

Getting taxonomic information

GBIF.taxon — Function

Get information about a taxon at any level

taxon(name::String)

This function will look for a taxon by its (scientific) name in the GBIF reference taxonomy.

Optional arguments are

rank::Union{Symbol,Nothing}=:SPECIES – the rank of the taxon you want. This is part of a controlled vocabulary, and can only be one of :DOMAIN, :CLASS, :CULTIVAR, :FAMILY, :FORM, :GENUS, :INFORMAL, :ORDER, :PHYLUM,, :SECTION, :SUBCLASS, :VARIETY, :TRIBE, :KINGDOM, :SUBFAMILY, :SUBFORM, :SUBGENUS, :SUBKINGDOM, :SUBORDER, :SUBPHYLUM, :SUBSECTION, :SUBSPECIES, :SUBTRIBE, :SUBVARIETY, :SUPERCLASS, :SUPERFAMILY, :SUPERORDER, and :SPECIES

strict::Bool=true – whether the match should be strict, or fuzzy

Finally, one can also specify other levels of the taxonomy, using kingdom, phylum, class, order, family, and genus, all of which can either be String or Nothing.

If a match is found, the result will be given as a GBIFTaxon. If not, this function will return nothing and give a warning.

Get information about a taxon at any level using taxonID

taxon(id::Int)

This function will look for a taxon by its taxonID in the GBIF reference taxonomy.

Searching for occurrence data

The most common task is to retrieve many occurrences according to a query. The core type of this package is GBIFRecord, which is a very lightweight type containing information about the query, and a list of GBIFRecord for every matching occurrence. Note that the GBIF "search" API is limited to 100000 results, and will not return more than this amount.

Single occurrence

GBIF.occurrence — Function

Return an interpreted occurrence given its key

occurrence(key::Union{String, Integer})

The key can be given as a string or as an integer.

As an example, we can retrieve the occurrence with the key 1258202889, using the following code:

using GBIF
occurrence(1258202889)

GBIF record 1258202889	Alfaroa	(Panama)

Multiple occurrences

GBIF.occurrences — Method

Retrieve latest occurrences based on a query

occurrences(query::Pair...)

This function will return the latest occurrences matching the queries – usually 20, but this is entirely determined by the server default page size. The query parametes must be given as pairs, and are optional. Omitting the query will return the latest recorded occurrences.

The arguments accepted as queries are the one documented on the GBIF API website.

GBIF.occurrences — Method

Retrieve latest occurrences for a taxon based on a query

occurrences(t::GBIFTaxon, query::Pair...)

Returns occurrences that correspond to a filter, given in q as a dictionary. When first called, this function will return the latest 20 hits (or whichever default page size GBIF uses). Future occurrences can be queried with next! or complete!.

When called with no arguments, this function will return a list of the latest 20 occurrences recorded in GBIF. Note that the GBIFRecords type, which is the return type of occurrences, implements the iteration interface. For example, this allows writing the following:

using GBIF
o = occurrences()
for single_occ in o
  print(single_occ.taxon.name)
end

Psittacus erithacusAsplenium radicansEugenia pyriformisMarasmius epiphyllusBerkleasmium moriformePseudovalsa umbonataHypoxylon cercidicolaNemania serpensHemitrichia intortaDiatrypella favaceaCryptocline taxicolaAsota heliconiaSelepaLemyra imparillisLimnonectes woodworthiLimnonectes woodworthiLimnonectes woodworthiPolystichum wawranumHymenophyllum atrovirensEpacris pauciflora

Query parameters

The queries must be given as pairs of

GBIF.occurrences — Method

Retrieve latest occurrences based on a query

occurrences(query::Pair...)

This function will return the latest occurrences matching the queries – usually 20, but this is entirely determined by the server default page size. The query parametes must be given as pairs, and are optional. Omitting the query will return the latest recorded occurrences.

The arguments accepted as queries are the one documented on the GBIF API website.

GBIF.occurrences — Method

Retrieve latest occurrences for a taxon based on a query

occurrences(t::GBIFTaxon, query::Pair...)

Returns occurrences that correspond to a filter, given in q as a dictionary. When first called, this function will return the latest 20 hits (or whichever default page size GBIF uses). Future occurrences can be queried with next! or complete!.

For example, we can get the data on observations of bats between -30 and 30 of latitude using the following syntax:

using GBIF
bats = GBIF.taxon("Chiroptera"; rank=:ORDER)
for occ in occurrences(bats, "decimalLatitude" => (-30.0, 30.0))
  println("$(occ.scientific) -- latitude = $(occ.latitude)")
end

Pteropus conspicillatus Gould, 1850 -- latitude = -16.798766
Pteropus giganteus (Brünnich, 1782) -- latitude = 24.543805
Cynopterus sphinx (Vahl, 1797) -- latitude = 22.252123
Rhynchonycteris naso (Wied-Neuwied, 1820) -- latitude = 16.368187
Pteropus giganteus (Brünnich, 1782) -- latitude = 27.165536
Rhynchonycteris naso (Wied-Neuwied, 1820) -- latitude = 10.605141
Pteropus giganteus (Brünnich, 1782) -- latitude = 18.92548
Hesperoptenus tickelli (Blyth, 1851) -- latitude = 22.121087
Pteropus conspicillatus Gould, 1850 -- latitude = -16.798661
Rhynchonycteris naso (Wied-Neuwied, 1820) -- latitude = 10.217022
Rhynchonycteris naso (Wied-Neuwied, 1820) -- latitude = 10.438246
Eptesicus fuscus (Palisot de Beauvois, 1796) -- latitude = 29.181554
Artibeus jamaicensis Leach, 1821 -- latitude = 20.654952
Artibeus jamaicensis Leach, 1821 -- latitude = 20.654952
Artibeus jamaicensis Leach, 1821 -- latitude = 20.654952
Lasiurus cinereus (Palisot de Beauvois, 1796) -- latitude = 19.414561
Cardioderma cor (Peters, 1872) -- latitude = -4.26467
Pteropus poliocephalus Temminck, 1825 -- latitude = -27.775096
Rhinolophidae -- latitude = -16.2
Chiroptera -- latitude = -16.2

Batch-download of occurrences

When calling occurrences, the list of possible GBIFRecord will be pre-allocated. Any subsequent call to occurrences! (on the GBIFRecords variable) will retrieve the next "page" of results, and add them to the collection:

GBIF.occurrences! — Function

Get the next page of results

This function will retrieve the next page of results. By default, it will walk through queries 20 at a time. This can be modified by changing the .query["limit"] value, to any value up to 300, which is the limit set by GBIF for the queries.

If filters have been applied to this query before, they will be removed to ensure that the previous and the new occurrences have the same status, but only for records that have already been retrieved.

using GBIF
can_most_recent = occurrences("hasCoordinate" => true, "country" => "CA")
occurrences!(can_most_recent)

20-element Array{GBIFRecord,1}:
 GBIF record 2542921773	Symplocarpus foetidus	(Canada)

 GBIF record 2542924359	Ligustrum vulgare	(Canada)

 GBIF record 2542924411	Calypte anna	(Canada)

 GBIF record 2542924486	Calidris alpina	(Canada)

 GBIF record 2542924491	Tamiasciurus hudsonicus	(Canada)

 GBIF record 2542926178	Anas platyrhynchos	(Canada)

 GBIF record 2542926212	Platanus occidentalis	(Canada)

 GBIF record 2542926214	Carpinus caroliniana	(Canada)

 GBIF record 2542926219	Polycauliona polycarpa	(Canada)

 GBIF record 2542927746	Betula alleghaniensis	(Canada)

 GBIF record 2542928882	Trifolium repens	(Canada)

 GBIF record 2542929843	Carya cordiformis	(Canada)

 GBIF record 2542929848	Acer spicatum	(Canada)

 GBIF record 2542929871	Dirca palustris	(Canada)

 GBIF record 2542930958	Spatula clypeata	(Canada)

 GBIF record 2542932037	Chelone glabra	(Canada)

 GBIF record 2542933035	Dryobates pubescens	(Canada)

 GBIF record 2542933059	Typha glauca	(Canada)

 GBIF record 2542938327	Betula alleghaniensis	(Canada)

 GBIF record 2542938415	Anas platyrhynchos	(Canada)