Filtering observations

Within a GBIFRecord object, the occurrences themselves are stored in a view. A view is basically an array that can be masked, so it is possible to retain all of the raw data, while only presenting the data that pass filtering.

Apply filters to the data

Base.filter! — Function

Filters a series of records

This function will take the filter function f and use it to mask the records that do not satisfy it. Note that if a record is already masked due to the application of a previous filter, its status will not be modified. The application of filters is therefore cumulative.

Note that while the usual filter! function removes objects, this one will only mask them. The objects can be unmasked with allrecords!. This design choice was made because GBIFRecords track the position in the API pages, and it seemed safer to keep all information. Note that when calling length or iterating over a GBIFRecords object, only the unmasked records will be shown.

It is possible to apply multiple filters at once, using the following syntax:

filter!.([f1, f2, f3], records)

Removing filters

GBIF.allrecords! — Function

Show all occurrences

This function reverses the action of filter!. It will unmask all records that have been removed under the current filters.

List of built-in filters

GBIF.have_both_coordinates — Function

Both latitude and longitude are present

This filter will only retain occurrences that have both a latitude and a longitude field.

GBIF.have_neither_zero_coordinates — Function

Neither latitude nor longitude are 0.0

This filter will only retain occurrences that are not at the (0,0) coordinate.

GBIF.have_no_zero_coordinates — Function

At most one of latitude and longitude is 0.0

This filter will only retain occurrences that have at most one coordinate being exactly 0.0.

GBIF.have_no_issues — Function

No known issues at all

This filter will retain no observation. At least not in theory, but it is very rare to have GBIF records with absolutely no issues. Its use is discouraged.

GBIF.have_ok_coordinates — Function

No coordinates issues except rounding and WGS84 assumption

This filter will retain observations that have no issues, with the exception of assuming WGS84 projection (which a large number of GBIF records do), and rounded coordinates (same thing). It is a reasonable filter for most use cases.

GBIF.have_a_date — Function

The date is not missing

This filter will remove observations with a missing date.

Making your own filters

Filter functions are all sharing the same declaration: they accept a single GBIFRecord object as input, and return a boolean as output. Think of the filter as a question you ask about the occurrence object. Does it have no know issues? If this is true, then we keep this record. If not, we reject it.

Filtering occurrences after download

The GBIFRecords objects can be used with the Query.jl package. For example, to get the observations from France in the most recent 20 observations, we can use:

using Query
o = occurrences()
@from i in o begin
    @where i.country == "France"
    @select {i.key, i.species}
    @collect
end