Occurrences

Retrieving occurrences

The most common task is to retrieve a number of occurrences. The core type of this package is GBIFRecord, which stores a number of data and metadata associated with observations of occurrences.

Getting a single occurrence

GBIF.occurrence — Function.

Return an interpreted occurrence given its key

The key can be given as a string or as an integer.

source

This can be used to retrieve [occurrence 1425976049][exocc], with

using GBIF
occurrence(1425976049)

[exocc]: https://www.gbif.org/occurrence/1425976049

Getting multiple occurrences

GBIF.occurrences — Method.

Search for occurrences

This function will return the latest occurrences – usually 20, but this is entirely determined by the server default page size. This is mostly useful to get a few results rapidly for illustration purposes.

source

When called with no arguments, this function will return a list of the latest 20 occurrences recorded in GBIF. Additional arguments can be specified to filter some occurrences. They are detailed in the "Using queries" section of this manual.

Note that the GBIFRecords type, returned by occurrences, implements all the necessary methods to iterate over collections. For example, this allows writing the following:

o = occurrences()
for single_occ in o
  println(o.taxonKey)
end

Batch-download of occurrences

GBIF.next! — Function.

Get the next page of results

This function will retrieve the next page of results. By default, it will walk through queries 20 at a time. This can be modified by changing the .query["limit"] value, to any value below 200.

source

GBIF.complete! — Function.

Get all pages of results

This function will retrieve all matches (up to the GBIF limit of 200000 records for a streaming query). It is recommended to set the limit to more than the default of 20 before calling this function. If not, this will trigger a lot of requests both from your end and on the GBIF infrastructure.

Internally, this function is simply calling next! until all records are exhausted.

source

Filtering occurrences after download

The GBIFRecords objects can be used with the Query.jl package. For example, to get the observations from France in the most recent 20 observations, we can use:

using Query
o = occurrences()
@from i in o begin
    @where i.country == "France"
    @select {i.key, i.species}
    @collect
end