Queries

Query parameters

GBIF.occurrences — Method.

Search for occurrences

Returns occurrences that correspond to a filter, given in q as a dictionary. When first called, this function will return the latest 20 hits (or whichever default page size GBIF uses). Future occurrences can be queried with next! or complete!.

source

Internal checks for query parameters

GBIF.check_records_parameters! — Method.

Checks that the queries for occurrences searches are well formatted

This is used internally.

Everything this function does is derived from the GBIF API documentation, including (and especially) the values for enum types. This modifies the queryset. Filters that are not allowed are removed, and filters that have incorrect values are dropped too.

This feels like the most conservative option – the user can always filter the results when they are returned.

source