Getting data

Retrieving data

Getting taxonomic information

GBIF.taxon — Function.

Get information about a taxon at any level

taxon(name::String)

This function will look for a taxon by its (scientific) name in the GBIF reference taxonomy.

Optional arguments are

rank::Union{Symbol,Nothing}=:SPECIES – the rank of the taxon you want. This is part of a controlled vocabulary, and can only be one of :DOMAIN, :CLASS, :CULTIVAR, :FAMILY, :FORM, :GENUS, :INFORMAL, :ORDER, :PHYLUM,, :SECTION, :SUBCLASS, :VARIETY, :TRIBE, :KINGDOM, :SUBFAMILY, :SUBFORM, :SUBGENUS, :SUBKINGDOM, :SUBORDER, :SUBPHYLUM, :SUBSECTION, :SUBSPECIES, :SUBTRIBE, :SUBVARIETY, :SUPERCLASS, :SUPERFAMILY, :SUPERORDER, and :SPECIES

strict::Bool=true – whether the match should be strict, or fuzzy

verbose::Bool=false – whether the query should print out information about the reply (not usually useful)

Finally, one can also specify other levels of the taxonomy, using kingdom, phylum, class, order, family, and genus, all of which can either be String or Nothing.

source

Getting occurrence data

The most common task is to retrieve a number of occurrences. The core type of this package is GBIFRecord, which stores a number of data and metadata associated with observations of occurrences.

Single occurrence

GBIF.occurrence — Function.

Return an interpreted occurrence given its key

occurrence(key::Union{String, Integer})

The key can be given as a string or as an integer.

source

This can be used to retrieve occurrence 1425976049, with

using GBIF
occurrence(1425976049)

Multiple occurrences

GBIF.occurrences — Method.

Retrieve latest occurrences

occurrences()

This function will return the latest occurrences – usually 20, but this is entirely determined by the server default page size. This is mostly useful to get a few results rapidly for illustration purposes.

source

GBIF.occurrences — Method.

Retrieve latest occurrences for a taxon

occurrences(t::GBIFTaxon)

Returns occurrences that correspond to a filter, given in q as a dictionary. When first called, this function will return the latest 20 hits (or whichever default page size GBIF uses). Future occurrences can be queried with next! or complete!.

source

When called with no arguments, this function will return a list of the latest 20 occurrences recorded in GBIF. Note that the GBIFRecords type, returned by occurrences, implements all the necessary methods to iterate over collections. For example, this allows writing the following:

o = occurrences()
for single_occ in o
  println(o.taxonKey)
end

Query parameters

GBIF.occurrences — Method.

Retrieve latest occurrences based on a query

occurrences(q::Dict)

Returns occurrences that correspond to a filter, given in q as a dictionary. When first called, this function will return the latest 20 hits (or whichever default page size GBIF uses). Future occurrences can be queried with next! or complete!.

source

GBIF.occurrences — Method.

Retrieve latest occurrences for a taxon based on a query

occurrences(t::GBIFTaxon, q::Dict)

Returns occurrences that correspond to a filter, given in q as a dictionary. When first called, this function will return the latest 20 hits (or whichever default page size GBIF uses). Future occurrences can be queried with next! or complete!.

source

Batch-download of occurrences

GBIF.next! — Function.

Get the next page of results

This function will retrieve the next page of results. By default, it will walk through queries 20 at a time. This can be modified by changing the .query["limit"] value, to any value below 200.

source

GBIF.complete! — Function.

Get all pages of results

This function will retrieve all matches (up to the GBIF limit of 200000 records for a streaming query). It is recommended to set the limit to more than the default of 20 before calling this function. If not, this will trigger a lot of requests both from your end and on the GBIF infrastructure.

Internally, this function is simply calling next! until all records are exhausted.

source