Retrieving data

Getting taxonomic information

GBIF.taxon — Function

Get information about a taxon at any level

taxon(name::String)

This function will look for a taxon by its (scientific) name in the GBIF reference taxonomy.

Optional arguments are

rank::Union{Symbol,Nothing}=:SPECIES – the rank of the taxon you want. This is part of a controlled vocabulary, and can only be one of :DOMAIN, :CLASS, :CULTIVAR, :FAMILY, :FORM, :GENUS, :INFORMAL, :ORDER, :PHYLUM,, :SECTION, :SUBCLASS, :VARIETY, :TRIBE, :KINGDOM, :SUBFAMILY, :SUBFORM, :SUBGENUS, :SUBKINGDOM, :SUBORDER, :SUBPHYLUM, :SUBSECTION, :SUBSPECIES, :SUBTRIBE, :SUBVARIETY, :SUPERCLASS, :SUPERFAMILY, :SUPERORDER, and :SPECIES

strict::Bool=true – whether the match should be strict, or fuzzy

Finally, one can also specify other levels of the taxonomy, using kingdom, phylum, class, order, family, and genus, all of which can either be String or Nothing.

If a match is found, the result will be given as a GBIFTaxon. If not, this function will return nothing and give a warning.

Get information about a taxon at any level using taxonID

taxon(id::Int)

This function will look for a taxon by its taxonID in the GBIF reference taxonomy.

Getting occurrence data

The most common task is to retrieve a number of occurrences. The core type of this package is GBIFRecord, which stores a number of data and metadata associated with observations of occurrences.

Single occurrence

GBIF.occurrence — Function

Return an interpreted occurrence given its key

occurrence(key::Union{String, Integer})

The key can be given as a string or as an integer.

This can be used to retrieve occurrence 1258202889, with

using GBIF
occurrence(1258202889)

GBIF record 1258202889	Alfaroa	(Panama)

Multiple occurrences

GBIF.occurrences — Method

Retrieve latest occurrences based on a query

occurrences(query::Pair...)

This function will return the latest occurrences matching the queries – usually 20, but this is entirely determined by the server default page size. The query parametes must be given as pairs, and are optional. Omitting the query will return the latest recorded occurrences.

GBIF.occurrences — Method

Retrieve latest occurrences for a taxon based on a query

occurrences(t::GBIFTaxon, query::Pair...)

Returns occurrences that correspond to a filter, given in q as a dictionary. When first called, this function will return the latest 20 hits (or whichever default page size GBIF uses). Future occurrences can be queried with next! or complete!.

When called with no arguments, this function will return a list of the latest 20 occurrences recorded in GBIF. Note that the GBIFRecords type, returned by occurrences, implements all the necessary methods to iterate over collections. For example, this allows writing the following:

using GBIF
o = occurrences()
for single_occ in o
  print(single_occ)
end

GBIF record 1039645472	Psittacus erithacus	(missing)
GBIF record 1086659096	Asplenium radicans	(Brazil)
GBIF record 1135437949	Berkleasmium moriforme	(Austria)
GBIF record 1135453319	Hypoxylon cercidicola	(Austria)
GBIF record 1135459704	Hemitrichia intorta	(Austria)
GBIF record 1830980264	Diatrypella favacea	(Austria)
GBIF record 1830984737	Cryptocline taxicola	(Austria)
GBIF record 1840974687	Asota heliconia	(Taiwan)
GBIF record 1840980692	Selepa	(Taiwan)
GBIF record 1841088166	Lemyra imparillis	(Taiwan)
GBIF record 1880819384	Limnonectes woodworthi	(Philippines)
GBIF record 1880819419	Limnonectes woodworthi	(Philippines)
GBIF record 1880819501	Limnonectes woodworthi	(Philippines)
GBIF record 1989841959	Psychidae	(Taiwan)
GBIF record 2003570740	Dendrocopos major	(Finland)
GBIF record 2003697052	Cyanistes caeruleus	(Finland)
GBIF record 2003776314	Periparus ater	(Finland)
GBIF record 2003823045	Parus major	(Finland)
GBIF record 2004038238	Emberiza citrinella	(Finland)
GBIF record 2004109769	Pica pica	(Finland)

Query parameters

GBIF.occurrences — Method

Retrieve latest occurrences based on a query

occurrences(query::Pair...)

This function will return the latest occurrences matching the queries – usually 20, but this is entirely determined by the server default page size. The query parametes must be given as pairs, and are optional. Omitting the query will return the latest recorded occurrences.

GBIF.occurrences — Method

Retrieve latest occurrences for a taxon based on a query

occurrences(t::GBIFTaxon, query::Pair...)

Returns occurrences that correspond to a filter, given in q as a dictionary. When first called, this function will return the latest 20 hits (or whichever default page size GBIF uses). Future occurrences can be queried with next! or complete!.

For example, we can get the data on observations of bats between -30 and 30 of latitudes using the following syntax:

using GBIF
bats = GBIF.taxon("Chiroptera"; strict=false)
for occ in occurrences(bats, "decimalLatitude" => (-30.0, 30.0))
  println("$(occ.scientific) -- latitude = $(occ.latitude)")
end

Pteropus conspicillatus Gould, 1850 -- latitude = -16.798766
Pteropus giganteus (Brünnich, 1782) -- latitude = 24.543805
Cynopterus sphinx (Vahl, 1797) -- latitude = 22.252123
Rhynchonycteris naso (Wied-Neuwied, 1820) -- latitude = 16.368187
Pteropus giganteus (Brünnich, 1782) -- latitude = 27.165536
Rhynchonycteris naso (Wied-Neuwied, 1820) -- latitude = 10.605141
Pteropus giganteus (Brünnich, 1782) -- latitude = 18.92548
Hesperoptenus tickelli (Blyth, 1851) -- latitude = 22.121087
Pteropus conspicillatus Gould, 1850 -- latitude = -16.798661
Rhynchonycteris naso (Wied-Neuwied, 1820) -- latitude = 10.217022
Rhynchonycteris naso (Wied-Neuwied, 1820) -- latitude = 10.438246
Eptesicus fuscus (Palisot de Beauvois, 1796) -- latitude = 29.181554
Artibeus jamaicensis Leach, 1821 -- latitude = 20.654952
Artibeus jamaicensis Leach, 1821 -- latitude = 20.654952
Artibeus jamaicensis Leach, 1821 -- latitude = 20.654952
Lasiurus cinereus (Palisot de Beauvois, 1796) -- latitude = 19.414561
Cardioderma cor (Peters, 1872) -- latitude = -4.26467
Pteropus poliocephalus Temminck, 1825 -- latitude = -27.775096
Rhinolophidae -- latitude = -16.2
Chiroptera -- latitude = -16.2

Batch-download of occurrences

GBIF.occurrences! — Function

Get the next page of results

This function will retrieve the next page of results. By default, it will walk through queries 20 at a time. This can be modified by changing the .query["limit"] value, to any value up to 300, which is the limit set by GBIF for the queries.

If filters have been applied to this query before, they will be removed to ensure that the previous and the new occurrences have the same status, but only for records that have already been retrieved.

using GBIF
can_most_recent = occurrences("hasCoordinate" => true, "country" => "CA")
occurrences!(can_most_recent)

20-element Array{GBIFRecord,1}:
 GBIF record 2542921773	Symplocarpus foetidus	(Canada)

 GBIF record 2542924359	Ligustrum vulgare	(Canada)

 GBIF record 2542924411	Calypte anna	(Canada)

 GBIF record 2542924486	Calidris alpina	(Canada)

 GBIF record 2542924491	Tamiasciurus hudsonicus	(Canada)

 GBIF record 2542926178	Anas platyrhynchos	(Canada)

 GBIF record 2542926212	Platanus occidentalis	(Canada)

 GBIF record 2542926214	Carpinus caroliniana	(Canada)

 GBIF record 2542926219	Polycauliona polycarpa	(Canada)

 GBIF record 2542927746	Betula alleghaniensis	(Canada)

 GBIF record 2542928882	Trifolium repens	(Canada)

 GBIF record 2542929843	Carya cordiformis	(Canada)

 GBIF record 2542929848	Acer spicatum	(Canada)

 GBIF record 2542929871	Dirca palustris	(Canada)

 GBIF record 2542930958	Spatula clypeata	(Canada)

 GBIF record 2542932037	Chelone glabra	(Canada)

 GBIF record 2542933035	Dryobates pubescens	(Canada)

 GBIF record 2542933059	Typha glauca	(Canada)

 GBIF record 2542938327	Betula alleghaniensis	(Canada)

 GBIF record 2542938415	Anas platyrhynchos	(Canada)