Filtering records

Filtering observations

Within a GBIFRecord object, the occurrences themselves are stored in a view. A view is basically an array that can be masked, so it is possible to retain all of the raw data, while only presenting the data that pass filtering.

Apply filtering to the data

GBIF.qualitycontrol! — Function.

Cleans a search output

This function loops through all records, and applies the filters to it. Filters are built-in or user-defined functions that return true when the record needs to be kept, and false when it needs to be discarded.

It is important to note that the records are not actually removed: they are masked from user view. This means that you can try different filtering strategies without having to re-query GBIF.

The optional filters argument is an array of functions, each of the functions must take a single GBIFRecord as an input, and return true or false.

source

List of filters

GBIF.have_both_coordinates — Function.

Both latitude and longitude are present

This filter will only retain occurrences that have both a latitude and a longitude field.

source

GBIF.have_neither_zero_coordinates — Function.

Neither latitude nor longitude are 0.0

This filter will only retain occurrences that are not at the (0,0) coordinate.

source

GBIF.have_no_zero_coordinates — Function.

At most one of latitude and longitude is 0.0

This filter will only retain occurrences that have at most one coordinate being exactly 0.0.

source

GBIF.have_no_issues — Function.

No known issues at all

This filter will retain no observation. At least not in theory, but it is very rare to have GBIF records with absolutely no issues. Its use is discouraged.

By design, it is the default argument for qualitycontrol! – it's your job to decide precisely which filters you need to use.

source

GBIF.have_ok_coordinates — Function.

No coordinates issues except rounding and WGS84 assumption

This filter will retain observations that have no issues, with the exception of assuming WGS84 projection (which a large number of GBIF records do), and rounded coordinates (same thing). It is a reasonable filter for most use cases.

source

Making your own filters

Filter functions are all sharing the same declaration: they accept a single GBIFRecord object as input, and return a boolean as output. Think of the filter as a question you ask about the occurrence object. Does it have no know issues? If this is true, then we keep this record. If not, we reject it.

Removing filters

GBIF.showall! — Function.

Show all occurrences

This function reverses the action of qualitycontrol!. It will unmask all records that have been removed under the current filters.

source

Filtering occurrences after download

The GBIFRecords objects can be used with the Query.jl package. For example, to get the observations from France in the most recent 20 observations, we can use:

using Query
o = occurrences()
@from i in o begin
    @where i.country == "France"
    @select {i.key, i.species}
    @collect
end